+++ lib-net.go
@@ -83,7 +83,7 @@ var throttle bool
 var WlogDisplay=true
 
 // Counting semaphore using a buffered channel
-func limitNumClients(f http.HandlerFunc, maxClients int, evalfs uint64) http.HandlerFunc {
+func limitNumClients(f http.HandlerFunc, maxClients int, evalfs uint32) http.HandlerFunc {
     sema := make(chan struct{}, maxClients)
     return func(w http.ResponseWriter, req *http.Request) {
         ctx:=req.Context()
@@ -239,7 +239,7 @@ type webstruct struct {
 
 func webRouter(w http.ResponseWriter, r *http.Request) {
 
-    evalfs,_:=r.Context().Value("evalfs").(uint64)
+    evalfs,_:=r.Context().Value("evalfs").(uint32)
 
     // we do not log by default
     // if globalvar log_web is true, then log to web_log_file (global)
@@ -648,7 +648,7 @@ func buildNetLib() {
     // error under all circumstances. We'll have track handles against ip/port here.
 
     slhelp["net_interfaces"] = LibHelp{in: "", out: "device_string", action: "newline separated list of device names."}
-    stdlib["net_interfaces"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["net_interfaces"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         i,_:=net.Interfaces()
         a:=""
         for _,v:=range i {
@@ -658,18 +658,18 @@ func buildNetLib() {
     }
 
     slhelp["web_display"] = LibHelp{in: "", out: "", action: "(debug)"}
-    stdlib["web_display"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_display"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         webRoutesAll()
         return nil,nil
     }
 
     slhelp["web_max_clients"] = LibHelp{in: "", out: "int", action: "(read-only) returns the maximum permitted client count for a web server."}
-    stdlib["web_max_clients"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_max_clients"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         return int(MAX_CLIENTS),nil
     }
 
     slhelp["web_serve_log"] = LibHelp{in: "args", out: "", action: "Write arguments to the web log file, if available."}
-    stdlib["web_serve_log"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_serve_log"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         switch len(args) {
         case 0:
             return nil,nil
@@ -686,7 +686,7 @@ func buildNetLib() {
                     "web_serve_decode() returns the following member fields:\n"+
                     "  .host (host:port), .method, .path, .remote (remote_ip:port) and .data (POST data).",
 }
-    stdlib["web_serve_decode"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_serve_decode"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==1 {
             switch args[0].(type) {
             case webstruct:
@@ -697,7 +697,7 @@ func buildNetLib() {
     }
 
     slhelp["web_serve_start"] = LibHelp{in: "docroot,port,vhost", out: "handle", action: "Returns an identifier for a new http server."}
-    stdlib["web_serve_start"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_serve_start"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 
         // validate args (docroot,port,addr)
 
@@ -805,7 +805,7 @@ func buildNetLib() {
     }
 
     slhelp["web_serve_stop"] = LibHelp{in: "handle", out: "success_flag", action: "Stops and discards a running http server."}
-    stdlib["web_serve_stop"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_serve_stop"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 
         var uid string
         if len(args)==1 {
@@ -828,7 +828,7 @@ func buildNetLib() {
     }
 
     slhelp["html_escape"] = LibHelp{in: "string", out: "string", action: "Converts HTML special characters to ampersand values."}
-    stdlib["html_escape"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["html_escape"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         var s string
         if len(args)==1 {
             switch args[0].(type) {
@@ -843,7 +843,7 @@ func buildNetLib() {
     }
 
     slhelp["html_unescape"] = LibHelp{in: "string", out: "string", action: "Converts a string containing ampersand values to include HTML special characters."}
-    stdlib["html_unescape"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["html_unescape"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         var s string
         if len(args)==1 {
             switch args[0].(type) {
@@ -859,7 +859,7 @@ func buildNetLib() {
 
 
     slhelp["web_serve_up"] = LibHelp{in: "handle", out: "bool", action: "Checks if a web server is still running."}
-    stdlib["web_serve_up"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_serve_up"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 
         var uid string
         if len(args)==1 {
@@ -881,7 +881,7 @@ func buildNetLib() {
     }
 
     slhelp["web_serve_path"] = LibHelp{in: "handle,action_type,request_regex,new_path", out: "string", action: "Provides a traffic routing instruction to a web server."}
-    stdlib["web_serve_path"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_serve_path"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=4 {
             return false,errors.New("bad call args in web_serve_path()")
         }
@@ -935,7 +935,7 @@ func buildNetLib() {
     // @todo: add a call for removing web_rules
 
     slhelp["web_serve_log_throttle"] = LibHelp{in: "start,freq", out: "", action: "Set the throttle controls for web server logging."}
-    stdlib["web_serve_log_throttle"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_serve_log_throttle"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=2 {
             return nil,errors.New("invalid args to web_serve_log_throttle()")
         }
@@ -954,13 +954,13 @@ func buildNetLib() {
 //  than the few seconds I've considered it for.
 
     slhelp["web_template"] = LibHelp{in: "handle,template_path", out: "processed_string", action: "Reads from either an absolute path or a docroot relative path (if handle not nil), with template instructions interpolated."}
-    stdlib["web_template"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_template"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
     return nil,errors.New("Not implemented.")
     }
 */
 
     slhelp["web_head"] = LibHelp{in: "loc_string", out: "bool", action: "Makes a HEAD request of the given [#i1]loc_string[#i0]. Returns true if retrieved successfully."}
-    stdlib["web_head"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_head"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=1 { return false,errors.New("Bad args (count) to web_head()") }
         _, down_code := head(args[0].(string))
         if down_code>299 {
@@ -970,7 +970,7 @@ func buildNetLib() {
     }
 
     slhelp["web_get"] = LibHelp{in: "loc_string", out: "structure", action: "Returns a [#i1]structure[#i0] with content downloaded from [#i1]loc_string[#i0]. [#i1].result[#i0] is the content string. [#i1].code[#i0] is the status code."}
-    stdlib["web_get"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_get"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=1 { return false,errors.New("Bad args (count) to web_get()") }
         s, down_code, _ := download(args[0].(string))
         if down_code>299 {
@@ -980,7 +980,7 @@ func buildNetLib() {
     }
 
     slhelp["web_custom"] = LibHelp{in: "method_string,loc_string[,[]assoc_headers_strings]", out: "string", action: "Returns a [#i1]string[#i0] with content downloaded from [#i1]loc_string[#i0]."}
-    stdlib["web_custom"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_custom"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         var method_string, loc_string string
         var headers =make(map[string]string)
 
@@ -1023,7 +1023,7 @@ func buildNetLib() {
     }
 
     slhelp["web_post"] = LibHelp{in: "loc_string,[]key_value_list", out: "result_string", action: "Perform a HTTP POST."}
-    stdlib["web_post"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_post"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=2 {
             return "",errors.New("invalid args to web_post()")
         }
@@ -1038,7 +1038,7 @@ func buildNetLib() {
     }
 
     slhelp["download"] = LibHelp{in: "url_string", out: "local_name", action: "Downloads from URL [#i1]url_string[#i0] and stores the returned data in the file [#i1]local_name[#i0]. Includes console feedback."}
-    stdlib["download"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["download"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if sf("%T",args[0])!="string" {
             return "",errors.New("Bad args (type) in download()")
         }
@@ -1049,7 +1049,7 @@ func buildNetLib() {
     }
 
     slhelp["web_download"] = LibHelp{in: "url_string,local_file", out: "bool_okay", action: "Downloads from URL [#i1]url_string[#i0] and stores the returned data in the file [#i1]local_file[#i0]."}
-    stdlib["web_download"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["web_download"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=2 { return false,errors.New("Bad args (count) to web_download()") }
         cont, down_code, header := download(args[0].(string))
         _=header
